Move the preassembly database env object into util.
